fix: Duplicated action id + refactoring
