+++ src/main/java/io/github/nahuel92/pit4u/action/PIT4UAction.java
@@ -28,7 +28,6 @@ import com.intellij.psi.search.GlobalSearchScopesCore.DirectoryScope;
 import io.github.nahuel92.pit4u.configuration.PIT4UConfigurationType;
 import io.github.nahuel92.pit4u.configuration.PIT4UEditorStatus;
 import io.github.nahuel92.pit4u.configuration.PIT4URunConfiguration;
-import io.github.nahuel92.pit4u.icon.PIT4UIcon;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.idea.maven.utils.MavenUtil;
 
@@ -38,10 +37,6 @@ import java.util.stream.Collectors;
 
 public final class PIT4UAction extends AnAction {
     private static final Logger LOG = Logger.getInstance(PIT4UAction.class);
-
-    PIT4UAction() {
-        getTemplatePresentation().setIcon(PIT4UIcon.ICON);
-    }
 
     private static boolean shouldShow(final AnActionEvent e) {
         final var project = e.getProject();
