+++ src/main/java/io/github/nahuel92/pit4u/highlighter/MutationDataService.java
@@ -8,15 +8,15 @@ import io.github.nahuel92.pit4u.highlighter.dto.Mutation;
 import org.jetbrains.annotations.NotNull;
 
 import java.util.Collection;
-import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
+import java.util.concurrent.ConcurrentHashMap;
 import java.util.stream.Collectors;
 
 @Service(Service.Level.PROJECT)
 public final class MutationDataService {
     private static final Logger LOG = Logger.getInstance(MutationDataService.class);
-    private final Map<String, List<Mutation>> mutationMap = new HashMap<>();
+    private final Map<String, List<Mutation>> mutationMap = new ConcurrentHashMap<>();
 
     MutationDataService() {
         LowMemoryWatcher.register(this::clear);
@@ -28,13 +28,12 @@ public final class MutationDataService {
 
     public void loadData(@NotNull final Collection<Mutation> mutations) {
         mutationMap.clear();
-        mutationMap.putAll(mutations.stream()
-                .collect(Collectors.groupingBy(Mutation::mutatedClass)));
+        mutationMap.putAll(mutations.stream().collect(Collectors.groupingBy(Mutation::mutatedClass)));
         LOG.debug("Mutation data loaded");
     }
 
     public Collection<Mutation> getMutationsForClass(final String fqName) {
-        return mutationMap.get(fqName);
+        return mutationMap.getOrDefault(fqName, List.of());
     }
 
     public boolean hasActiveMutations() {
